requirements.txt
PySide6
mysql-connector-python
requests
xmltodict
Pillow
pandas
openpyxl